+++ relays/ethereum/src/main.rs
@@ -27,7 +27,9 @@ mod exchange_loop;
 mod exchange_loop_metrics;
 mod headers;
 mod instances;
+mod message_lane;
 mod message_lane_loop;
+mod message_delivery_race;
 mod metrics;
 mod rpc;
 mod rpc_errors;
